+++ stroom-processor/stroom-processor-impl-db/src/main/java/stroom/processor/impl/db/RecordToProcessorFilterTrackerMapper.java
@@ -24,7 +24,9 @@ class RecordToProcessorFilterTrackerMapper implements Function<Record, Processor
         processorFilterTracker.setLastPollMs(record.get(PROCESSOR_FILTER_TRACKER.LAST_POLL_MS));
         processorFilterTracker.setLastPollTaskCount(record.get(PROCESSOR_FILTER_TRACKER.LAST_POLL_TASK_COUNT));
         processorFilterTracker.setStatus(ProcessorFilterTrackerStatus.PRIMITIVE_VALUE_CONVERTER
-                .fromPrimitiveValue(record.get(PROCESSOR_FILTER_TRACKER.STATUS)));
+                .fromPrimitiveValue(
+                        record.get(PROCESSOR_FILTER_TRACKER.STATUS),
+                        ProcessorFilterTrackerStatus.CREATED));
         processorFilterTracker.setMessage(record.get(PROCESSOR_FILTER_TRACKER.MESSAGE));
         processorFilterTracker.setMetaCount(record.get(PROCESSOR_FILTER_TRACKER.META_COUNT));
         processorFilterTracker.setEventCount(record.get(PROCESSOR_FILTER_TRACKER.EVENT_COUNT));