+++ CHANGELOG.md
@@ -12,6 +12,8 @@ DO NOT ADD CHANGES HERE - ADD THEM USING log_change.sh
 ~~~
 
 
+* Issue **#3255** : Fix error when creating a new processor filter.
+
 * Issue **#3204** : More changes to improve task creation performance.
 
 * Issue **#3204** : Split UNPROCESSED(1) task state into CREATED(0) and QUEUED(1) to improve task creation performance and management.